Add files via upload
VOICE TO TEXT  
"To make a new "Captioning Font" in Google WebFonts, that will work in the "Captioning Settings" with "Lyrics and Speech Visualization",  I had to write 2405 lines of code in C.  And this is just the beginning." -Graeme Kilshaw  

LYRICS & SPEECH VISUALIZATION
"Here is the FriendshipCube Google WebFont Stack for Lyrics & Speech Visualization... live... in real-time." -Graeme Kilshaw, Software Engineer 

Our current goal is to demo our unique captioning settings with Google WebFonts... and to share our Lyrics and Speech Visualization Tool with the world.  This "Captioning WebFont Software" contributes to accessibility, inclusivity, and interoperability.  Here is a new portal to upload speech to text in real-time, and caption lyrics or speech, live.  This software contributes to peace and security, via databases that operate with neural networks and machine learning from the edge to the core to the cloud.) 

CAPTIONING FONT

#ifndef CAPTIONING FONT
#define CAPTIONING FONT 1 

Google WebFonts 

/*-----------------
 *  PUBLIC FONT
 *----------------*/
/*--------------------
 *  ALL CUSTOM DATA
 *--------------------*/
/*---------------------
 *  CHARACTER MAPPING
 *--------------------*/
 /*---------------------
 *  GLYPH DESCRIPTION
 *--------------------*/
/*-----------------
 *    BITMAPS
 *----------------*/
#endif /*#if CAPTIONING FONT*/
*/

https://github.com/FriendshipCube/ttf/blob/main/Captioning%20Font.c